+++ .prettierrc.json
@@ -3,5 +3,5 @@
   "endOfLine": "auto",
   "printWidth": 120,
   "semi": false,
-  "singleQuote": true,
+  "singleQuote": true
 }
