--- src/components/player/types/index.ts
@@ -1,1 +1,0 @@
-export * from './Track'
